requirements.txt
pandas
seaborn
matplotlib
aiohttp
python-dotenv
aiofiles
streamlit
loguru